add explicit order by for latest account (#5669)
closes: https://github.com/spacemeshos/go-spacemesh/issues/4825

it was overlooked in initial development. the order is not guaranteed by sqlite despite primary key ordered by in the order that is expected by vm . so sqlite free to choose any order it "likes" for optimal reasons. generally the optimal should be iterating over (account, layer_update) tree, but on different environment it may not be true