Fix possible race in ProtocolDriver on Close (#5707)
## Motivation

We have recently seen a few reports of this issue. While this isn't the actual cause of any problems (as something else needs to go wrong first for this problem to even trigger), it still makes it hard to identify the real cause of a problem.

This fixes the race so we can more easily identify the actual problem of a node shutting down early.