systest: fix RBAC handling (#5732)
## Motivation

It may be convenient to run the systests against an existing cluster, yet the current setup requires admin permissions for the default service account. Applying such settings compromises cluster security and should only be done on throwaway clusters. It's also not considered to be a good practice in general.

Besides, `waitPod` was sometimes crashing due to bad type cast.