increase tortoise window for past layers on mainnet (#5762)
closes: https://github.com/spacemeshos/go-spacemesh/issues/5659

in order to decode ballots while syncing layers before 30_000 window needs to be set to 10_000 (this is what this value was originally). otherwise many ballots can't be decoded as they reference base ballots outside of 4032 window. this breaks the chain of votes and make tortoise unable to cross necessary weight.

this change introduces HistoricalWindowSize parameter in tortoise, that sets window size to 10_000 for layers before 30_000.
it also removes eviction logic from atxsdata and moves it inside tortoise, so that we don't have to have similar conditionals in more than 1 place.