merge-nodes: Clarify that local.key exists in "appears supervised" errors
The current "appears to be a supervised node" error messages show an
assumption about the cause of the error, but don't show what this
assumption is based on.

This error message has lead to multiple users expressing their confusion
on the official Discord. Indicating the actual cause in the error
message hopefully reduces that confusion a bit.