merge-nodes: Clarify that local.key exists in "appears supervised" errors (#5781)
The current "appears to be a supervised node" error messages show an assumption about the cause of the error, but don't show what this assumption is based on.

This error message has led to multiple users expressing confusion on the official Discord. Indicating the actual cause in the error message hopefully slightly reduces that confusion.