Support poet certificates (#5221)
## Motivation
Closes #5204 

## Changes
Added support for registering in the poet services using a certificate instead of PoW. It still supports both ways and will fallback to the PoW if:
- certificate is not available (couldn't certify)
- the poet service doesn't support certificates

There is a new interface in `activation` package `certifierService` that provides a generic way to obtain a certificate for a given poet. It abstracts away the fact that a POST proof is used - it is hidden in `certifierServiceClient`, which could in future have different implementations for other poet services.

Changed the `initial_post` table to `post` as it now holds the latest POST proof. Relying solely on initial post proof OR an existing ATX could not work in case of a checkpoint (in fact this way it failed systests).

## Test Plan
- added UTs
- added system tests