p2p: server: close streams upon context cancellation (#5981)
## Motivation

When a P2P Server request is canceled or a server is stopped while serving a request, some code paths that involve blocking I/O calls may hang b/c the P2P stream is not closed upon cancellation of the context that was passed to `StreamRequest` / `Request` or `Run`. On top of that, the context which is passed to server handlers is not canceled when the server exits.

This was split off #5769 where a P2P test was hanging when trying to stop the P2P servers.