Copy out atx blob from the db query cache (#5993)
## Motivation

Sharing contents of the SQL query cache is risky as it must not be modified by the users. There are signs that it might be modified somewhere (cached ATX blobs are invalid), but I couldn't track where exactly. In this case, it's safer to always copy-out the bytes to avoid unintended modifications.