api: add hole punching info and data transfer stats (#5792)
## Motivation

Current `AdminService.PeerInfoStream` doesn't provide enough information about whether hole punching was used while connecting to a particular peer, and also doesn't reflect actual utilization of each peer, which makes it harder to see whether the peer connections are being used efficiently.
Also, importantly, in case if `go-spacemesh` causes network overload it may be hard to find out what kind of traffic is causing the problems. It would be nice to have some overview of the traffic even in case if the metrics are disabled. The current data transfer rates should be easy to view even without any Prometheus to do `rate` calculations.