checkpoint: add --ignore-checkpoint-req-errors
When this flag is specified, go-spacemesh doesn't exit when recovery
URI request fails. Retries are still done, though.
This is needed b/c in the context of TestRecovery systest, all the
nodes are created with --recovery-uri flag passed, including the
bootnodes. This way, they can't successfully start unless the
bootstrapper is already running. But the bootstrapper requires an
working go-spacemesh node to start up, thus we get a circular
dependency.