Malfeasance proof db update (#6378)
## Motivation

This separates the persistence for malfeasance proofs based on v1 and v2.

For v1 proofs will stay in the `identities` table and the code storing and fetching those proofs has been simplified. For v2 information about marriages moves to the new `marriages` table, and malfeasance proofs into the new `malfeasance` table.