sync2: improve rangesync Receive semantics
Given that after recent item sync is done (if it's needed at all), the
range set reconciliation algorithm no longer depends on newly received
item being added to the set, we can save memory by not adding the
received items during reconciliation.

During real sync, the received items will be sent to the respective
handlers and after the corresponding data are fetched and validated,
they will be added to the database, without the need to add them to
cloned OrderedSets which are used to sync against particular peers.