Extend TestSmeshing to also deploy an older version of go-spacmesh (#6403)
## Motivation

To detect errors that would break compatibility with previous versions of the node I extended `TestSmeshing` to deploy 25% of the nodes in the test using an older version of go-spacemesh. So if any change changes protocol rules that would break consensus like the error fixed in https://github.com/spacemeshos/go-spacemesh/pull/6398 or the error fixed in https://github.com/spacemeshos/go-spacemesh/pull/5839.