p2p: server: fix incorrect bufio usage
It is not correct to use `bufio.Buffer` to read the initial request
and then passing the underlying stream to the `StreamRequest`
callback, as `bufio.Buffer` may happen to read more data than
necessary, making it unavailable for the `StreamRequest` callback.
This behavior has been obvserved when using QUIC, which tends to
coalesce multiple writes more often.