sync2: only handle synced keys during commit
It turned out that sync interactions are happening rather quickly,
and thus it is not really practical to try and begin handling arriving
keys (e.g. ATX IDs) during sync itself. Moreover, on-the-fly key
handling was actually only used to register peer IDs for each ATX ID
to fetch the actual blob from, and that can be done in the handler's
Commit() method just as well.