sync2: fptree: don't get next in FingerprintInterval if not needed
Retrieving the next item when fingerprinting an interval is only
useful during slow splitting, when we need to split a range but "easy
split" (involving only in-memory FPTree) fails. In other cases,
the next item is not really needed and getting it may incur an
unwanted database access.

This also fixes intermittent failures of TestFPTreeManyItems test.