sync2: eliminate unneeded DB accesses when peers are in sync
`FPTree`, and thus `DBSet` hold enough information in memory for peers
to decide whether they are in sync or not. In fully synced state, the
peers only exchange probe requests that cover the whole set, and we
only need to know fingerprint and count of the set. For that, no
database queries are necessary.

This change eliminates unnecessary database access, simplifies
`OrderedSet` interface and adds a test case that checks that synced
peers do not make any queries except periodic queries needed to
advance `DBSet` (include any newly added items).