sql: reduce syncv2 conn pool usage and add pool utilization metrics
Database.WithConnection is always used by syncv2 for any probe/sync
operations against peers, but the actual connection it acquires from
the pool is not actually used sometimes, when all the necessary
information can be retrieved from the FPTree. Also, while holding the
connection for the duration of sync/probe works best in absolute
majority of cases, in case of a very slow peer it may cause unwanted
extra utilization of the connection pool.

Another problem is that there's currently no metrics for the actual
database connection pool utilization, and connection pool wait latency
metrics for the main state database pool, api pool and the local
database are not collected separately.

This change makes `Database.WithConnection` acquire the connection
from the pool lazily, that is, upon the first query, and release it
back if there are no queries executed for certain amount of
time (defaults to 10ms), which will prevent slow syncv2 peers from
holding connections for too long. In the latter case, the connection
is re-acquired from the pool upon the following query.

Also, this change adds connection pool utilization metric, and
separates the connection wait latency metrics and pool utilization
metrics for th main state database pool, api pool and the local
database.